refactor: moves ComboBox logix to custom hooks
